third_party/WebKit/LayoutTests/inspector/console/console-log-linkify-stack-in-errors-expected.txt

CONSOLE MESSAGE: line 5: Error: Some test
    at namedFunction (foob.js:5:17)
    at foob.js:8:1
CONSOLE MESSAGE: line 11: Error: line
break
    at forStack (console-log-linkify-stack-in-errors.html:14:1
CONSOLE MESSAGE: line 30: TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-stack-in-errors.html:34:1
CONSOLE MESSAGE: line 36: Error message without stacks http://www.chromium.org/
CONSOLE MESSAGE: line 38: Error valid stack #2
    at http://www.chromium.org/boo.js:40:70
    at foo(http://www.chromium.org/foo.js:10:50)
CONSOLE MESSAGE: line 39: Error valid stack #3
    at http://www.chromium.org/foo.js:40
CONSOLE MESSAGE: line 20: ReferenceError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:41:8
CONSOLE MESSAGE: line 20: EvalError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:42:8
CONSOLE MESSAGE: line 20: SyntaxError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:43:8
CONSOLE MESSAGE: line 20: RangeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:44:8
CONSOLE MESSAGE: line 20: TypeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:45:8
CONSOLE MESSAGE: line 20: URIError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:46:8
CONSOLE MESSAGE: line 48: Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70
CONSOLE MESSAGE: line 49: Error broken stack #2
    at function_name(foob.js:20:30
CONSOLE MESSAGE: line 50: Error broken stack #3
    at function_name(foob:20.js:30   bla
CONSOLE MESSAGE: line 51: Error broken stack #4
    at function_name)foob.js:20:30(
CONSOLE MESSAGE: line 52: Error broken stack #5
    at function_name foob.js:20:30)
CONSOLE MESSAGE: line 53: Error broken stack #6
    at foob.js foob.js:40:70
CONSOLE MESSAGE: line 9: Error
    at Failure.letsFailWithStack (stack-with-sourceMap.js:19:20)
    at <anonymous>:1:1
    at Object.InjectedScript._evaluateOn (<anonymous>:777:140)
    at Object.InjectedScript._evaluateAndWrap (<anonymous>:710:34)
    at Object.InjectedScript.evaluate (<anonymous>:626:21)
Test that console.log(new Error().stack) would linkify links in stacks for sourceUrls and sourceMaps Bug 424001.

foob.js:5 Error: Some test
    at namedFunction (foob.js:5)
    at foob.js:8 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:11 Error: line
break
    at forStack (console-log-linkify-stack-in-errors.html:11)
    at console-log-linkify-stack-in-errors.html:14 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:30 TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-stack-in-errors.html:28)
    at console-log-linkify-stack-in-errors.html:34 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:36 Error message without stacks http://www.chromium.org/ console-message-text source-code > console-message-url webkit-html-resource-link > webkit-html-external-link
console-log-linkify-stack-in-errors.html:38 Error valid stack #2
    at www.chromium.org/boo.js:40
    at foo(www.chromium.org/foo.js:10) console-message-text source-code > console-message-url webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:39 Error valid stack #3
    at www.chromium.org/foo.js:40 console-message-text source-code > console-message-url webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 ReferenceError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:41 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 EvalError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:42 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 SyntaxError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:43 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 RangeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:44 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 TypeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:45 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:20 URIError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20)
    at stack1 (console-log-linkify-stack-in-errors.html:22)
    at console-log-linkify-stack-in-errors.html:46 console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
console-log-linkify-stack-in-errors.html:48 Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70 console-message-text source-code > console-message-url webkit-html-resource-link
console-log-linkify-stack-in-errors.html:49 Error broken stack #2
    at function_name(foob.js:20:30 console-message-text source-code > console-message-url webkit-html-resource-link
console-log-linkify-stack-in-errors.html:50 Error broken stack #3
    at function_name(foob:20.js:30   bla console-message-text source-code > console-message-url webkit-html-resource-link
console-log-linkify-stack-in-errors.html:51 Error broken stack #4
    at function_name)foob.js:20:30( console-message-text source-code > console-message-url webkit-html-resource-link
console-log-linkify-stack-in-errors.html:52 Error broken stack #5
    at function_name foob.js:20:30) console-message-text source-code > console-message-url webkit-html-resource-link
console-log-linkify-stack-in-errors.html:53 Error broken stack #6
    at foob.js foob.js:40:70 console-message-text source-code > console-message-url webkit-html-resource-link
stack-with-sourceMap.coffee:3 Error
    at Failure.letsFailWithStack (stack-with-sourceMap.coffee:3)
    at window.failure (stack-with-sourceMap.coffee:7)
    at <anonymous>:1:1
    at Object.InjectedScript._evaluateOn (<anonymous>:777:140)
    at Object.InjectedScript._evaluateAndWrap (<anonymous>:710:34)
    at Object.InjectedScript.evaluate (<anonymous>:626:21) console-message-text source-code > console-message-url webkit-html-resource-link >  webkit-html-resource-link >  webkit-html-resource-link
